--- a/schema/sqitch.plan
+++ b/schema/sqitch.plan
@@ -364,3 +364,6 @@ shipping/views [shipping/views@2021-08-05] 2021-08-10T03:53:29Z Kristen Schwabe-
 
 shipping/views [shipping/views@2021-08-10] 2021-08-23T23:57:21Z Kristen Schwabe-Fry <2600836+kschwabefry@users.noreply.github.com> # Add tiny swabs collection to return of results shipping views
 @2021-08-23 2021-08-24T00:05:31Z Kristen Schwabe-Fry <2600836+kschwabefry@users.noreply.github.com> # Schema as of 23 August 2021
+
+shipping/views [shipping/views@2021-08-23] 2021-09-07T19:53:05Z Todd Seidelmann <seidelma@uw.edu> # Add new HCT REDCap URL
+@2021-09-07 2021-09-07T19:54:12Z Todd Seidelmann <seidelma@uw.edu> # Schema as of 7 September 2021