This PR completely cuts the affected views over to the new REDCap server. It shouldn't be deployed to prod until the new server is actually getting fed with prod data.