@seidelma 2021-08-18
I rebased with the changes Kristen suggested--I hope that simply checking for both Redcap projects is a sane way to do it.