@seidelma 2021-08-19
I've done some testing with `EXPLAIN` on these view-definition queries to make sure that using `WHERE .. OR ..` to select multiple Redcap servers isn't going to blow up like we saw with `ARRAY`. 

There's a significant performance penalty due to the BitmapOr/Bitmap Heap Scan operation, such that queries with `WHERE .. OR ..` are many times slower:

shipping.uw_reopening_encounters_v1:
  old: Execution time: 3.744 ms
  new: Execution time: 10432.748 ms

shipping.__uw_encounters:
  old: Execution time: 4.864 ms
  new: Execution time: 1477.948 ms

I guess the question is, is this an acceptable performance penalty?