@kschwabefry 2021-08-20
@seidelma good thinking ahead on looking at performance. Can you share the explain output?