@seidelma 2021-08-20
I've done some testing with a few different strategies to deal with this performance hit.

First, using the JSON 'get' operator (`->`) to reach into the `encounter.details` field and pull out the `url` tag. This is actually much slower than the 'contains' (`@>`) operator, because Postgres can't use the index for that, and it's got to do a full table scan:

    explain analyze verbose select encounter_id as enrollment_encounter_id
      , individual_id as enrollment_individual_id
      from warehouse.encounter
      where
          encounter.details @> '{"_provenance": {"redcap": {"url": "https://hct.redcap.rit.uw.edu/", "project_id":45, "event_name": "encounter_arm_1"}}}';
    Execution time: 1.315 ms

    explain analyze verbose select encounter_id as enrollment_encounter_id
      , individual_id as enrollment_individual_id
      from warehouse.encounter
      where
           encounter.details->'_provenance'->'redcap'->>'url' = 'https://hct.redcap.rit.uw.edu/';
     Execution time: 4560.658 ms

Next, to avoid the use of `WHERE .. OR ..` through `UNION ALL`, to basically concatenate the two queries (one for each Redcap server) together. Performance for this is similar to `WHERE .. OR ..`, because the query planner reorganizes it into the same parallel bitmap heap scan.

Failing that, I tried @davereinhart's solution to troubleshooting slow queries the other day: `encounter.details @> ANY( VALUES(..))`. That doesn't seem to work here--the query ends up being much, much slower than any of the other stuff I've done because it fails to use the index, and instead falls back to a seq scan on the encounter table.

Then I thought about reorganizing things into subqueries so we can avoid the JSON and look at integers using `CASE` to assign a number to each server. Unfortunately that also results in a seq scan.

Haven't given up yet, but that's where I'm currently at.